move the enum classes inside the entities its belong to
