+++ src/main/kotlin/data/audit/datasource/mongo/MongoAudit.kt
@@ -19,5 +19,4 @@ class MongoAudit(
         return auditCollection.find(filter).toList()
     }
 
-
 }
