+++ src/main/kotlin/data/task/mapper/TaskMapper.kt
@@ -4,7 +4,7 @@ import data.task.model.TaskDto
 import logic.model.Task
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
-import java.util.*
+import java.util.UUID
 
 class TaskMapper {
     private val dateFormatter = DateTimeFormatter.ISO_DATE
