+++ src/main/kotlin/logic/authentication/validation/PasswordValidator.kt
@@ -1,4 +1,4 @@
-package logic.authentication.validtion
+package logic.authentication.validation
 
 interface PasswordValidator {
     fun validatePassword(password: String)
