--- src/main/kotlin/logic/model/EntityType.kt
@@ -1,5 +1,0 @@
-package logic.model
-
-enum class EntityType {
-    PROJECT, TASK
-}
