--- src/main/kotlin/logic/model/UserType.kt
@@ -1,5 +1,0 @@
-package logic.model
-
-enum class UserType {
-    ADMIN, MATE
-}
