+++ src/main/kotlin/presentation/project/EditStateOfProjectUI.kt
@@ -2,7 +2,7 @@ package presentation.project
 
 import logic.audit.CreateAuditUseCase
 import logic.model.Audit
-import logic.model.EntityType
+import logic.model.Audit.EntityType
 import logic.project.EditStateOfProjectUseCase
 import logic.project.GetAllProjectsUseCase
 import logic.user.GetLoggedUserUseCase
