+++ src/test/kotlin/data/authentication/datasource/CsvAuthenticationTest.kt
@@ -10,7 +10,7 @@ import io.mockk.just
 import io.mockk.mockk
 import io.mockk.verify
 import logic.model.User
-import logic.model.UserType
+import logic.model.User.UserType
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.util.UUID
