+++ src/test/kotlin/data/authentication/datasource/CsvUserParserTest.kt
@@ -4,7 +4,7 @@ import com.google.common.truth.Truth.assertThat
 import data.authentication.datasource.csv.CsvUserParser
 import logic.exceptions.CanNotParseUserException
 import logic.model.User
-import logic.model.UserType
+import logic.model.User.UserType
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
